got halfway to build

+++ common/database.cpp
@@ -738,7 +738,7 @@ bool Database::LoadVariables()
 
 	LogInfo(
 		"Loaded [{}] Variable{}",
-		Strings::Commify(l.size()),
+		Strings::Commify(static_cast<uint64>(l.size())),
 		l.size() != 1 ? "s" : ""
 	);
 
@@ -2036,7 +2036,7 @@ bool Database::CopyCharacter(
 			size_t rows_copied = insert_rows.size(); // Rows copied for this table
 			total_rows_copied += rows_copied; // Increment grand total
 
-			LogInfo("Copying table [{}] rows [{}]", table_name, Strings::Commify(rows_copied));
+			LogInfo("Copying table [{}] rows [{}]", table_name, Strings::Commify(static_cast<uint64>(rows_copied)));
 
 			if (!insert.ErrorMessage().empty()) {
 				LogError("Error copying table [{}] [{}]", table_name, insert.ErrorMessage());
@@ -2056,7 +2056,7 @@ bool Database::CopyCharacter(
 		"Character [{}] copied to [{}] total rows [{}]",
 		source_character_name,
 		destination_character_name,
-		Strings::Commify(total_rows_copied)
+		Strings::Commify(static_cast<uint64>(total_rows_copied))
 	);
 
 	return true;
